net: dsa: VLAN-based port separation for switches that don't support tagging
Current open issue: only works when switch ports are standalone, and
when added to a VLAN-unaware bridge. When added to a bridge that has
vlan_filtering 1, the bridge core will install its own VLANs and reset
the pvids through switchdev. For the bridge core, switchdev is a
write-only pipe. All VLAN-related state is kept in the bridge core and
nothing is read from DSA/switchdev or from the driver. So the bridge
core will break this port separation because it will install the
vlan_default_pvid into all switchdev ports.

I am thinking about refactoring this newly added tagging_vid_from_port
(by me) dsa_switch_ops into a net_device_ops like ndo_vlan_default_pvid
or similar. The bridge core would see that its slave netdevice supports
this operation, and instead of installing its own vlan_default_pvid on
this slave netdevice, it would install a pvid that the driver ensures is
unique (and not the default which is 1).

The challenge I see is that in the DSA rcv callback, I would have to
perform an iterative reverse lookup to find the correct switch port
based on ndo_vlan_default_pvid. I can't use dsa_slave_to_port because
with a single translation function (port -> pvid) and none in the
reverse direction (pvid -> port) as I have now (the dsa_switch_ops
tagging_vid_from_port), I don't know what the numeric value of the port
is. And even if I knew, the port is a bridge slave, and its pvid is
subject to user configuration. How would I ensure that the user doesn't
reset the pvid to a different value, or to a non-unique value within
this DSA switch tree?

Or I could fend this issue by not adding the ndo_vlan_default_pvid
callback, but instead add a bridge core property along the lines of
vlan_unique_pvids. When this is user-requested, the bridge core would
install a pvid equal to the slave's netdevice ifindex, which is
guaranteed unique, and a small number. No back-and-forth translation.

But let's not forget that DSA fundamentally needs to add unique pvids to
switch ports even when they're not joined to a vlan-aware bridge. So the
solution would need to be generic enough that DSA could use it as well.
Also because of current design of the bridge core, DSA and bridge can't
communicate, they just need to follow the same reference.

That reference can be a new netdev feature bit added to its
vlan_features, something along the lines of NETIF_F_HW_VLAN_PORT_SEPARATION.
When the driver declares that feature, both DSA and bridge core take
care of installing a pvid on each switch port that is numerically equal
to its ifindex.

The final challenge is that not all switch ports are equal in this
setup. The CPU port needs to transmit tagged packets (VLAN trunk) in
order for the DSA rcv code to be able to decode source information.
But the bridge code has absolutely no idea which switch port is the CPU
port, if nothing else then just because there is no netdevice registered
by DSA for the CPU port.

So the bridge core cannot fully handle this anyway. DSA would need to
get smart during bridge_join, and if the master bridge is
vlan_filtering, DSA would need to do some additional fixups and make the
egress of the CPU port tagged. At this point I do wonder how does DSA
allow the user to specify that they want the CPU port to do VLAN
trunking anyway.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>