[OUT] net: dsa: sja1105: Don't advertise Ethernet flow control
The switch MAC does not consume, nor does it emit pause frames. It
simply forwards them as any other Ethernet frame (and since the DMAC is,
per IEEE spec, 01-80-C2-00-00-01, it means they are filtered as
link-local traffic and forwarded to the CPU, which can't do anything
useful with them).

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>