ARM: dts: ls1021: Use different compatible for MDIO node
I know it is not your fault to use this compatible string as it already
used this for mdio node in the device tree, but it is somewhat ambiguous
to use the same compatible string as the ethernet controller.  And this
dts is the only one in all kernel device trees using this compatible
string for mdio node.  I would think it will be more clear to use
"fsl,etsec2-mdio" or "fsl,etsec2-tbi".
^
Leo Li

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>