ether: Add dedicated Ethertype for pseudo-802.1Q DSA tagging
With the newly added DSA_TAG_PROTO_8021Q, switch devices that don't
support a native insertion/extraction header on the CPU port may still
enjoy the benefits of port isolation with a custom VLAN tag.

For this, they need to have a customizable TPID in hardware and a new
Ethertype to distinguish between real 802.1Q traffic and the private
tags used for port separation.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>