net: dsa: Allow drivers to modulate between presence and absence of tagging
Frames get processed by DSA and redirected to switch port net devices
based on the ETH_P_XDSA multiplexed packet_type handler found by the
network stack when calling eth_type_trans().

The running assumption is that once the DSA .rcv function is called, DSA
is always able to decode the switch tag in order to change the skb->dev
from its master.

However there are tagging protocols (such as the new
DSA_TAG_PROTO_SJA1105) where this assumption is not completely true,
since switch tagging piggybacks on the absence of a vlan_filtering
bridge.

Having DSA receive untagged traffic would put it in an impossible
situation: the eth_type_trans() function would invoke the DSA .rcv(),
which could not change skb->dev, then eth_type_trans() would be invoked
again, which again would call the DSA .rcv, and the packet would never
be able to exit the DSA filter and would spiral in a loop until the
whole system dies.

This happens because eth_type_trans() doesn't actually look at the skb
(so as to identify a potential tag) when it deems it as being
ETH_P_XDSA. It just checks whether skb->dev has a DSA private pointer
installed (therefore it's a DSA master) and that there exists a .rcv
callback (everybody except DSA_TAG_PROTO_NONE has that). This is
understandable as there are many switch tags out there, and exhaustively
checking for all of them is far from ideal.

The solution lies in the observation that a more nuanced check can be
made when eth_type_trans() determines that switch tagging is used or
not. In a way, this reverts patch "717ffbfb28ac net: dsa: remove
dsa_uses_tagged_protocol", but instead of adding it back as a DSA
function, it is now a boolean property. This is because the driver might
actually know better when it can and can't support switch tagging.

With this patch, all tagging protocols can morph at runtime into the
DSA_TAG_PROTO_NONE on receive, by setting cpu_dp->uses_tag_protocol = 0.
This permits them to at least terminate traffic through the master net
device. Their .rcv callback no longer even gets called in this mode.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>