net: dsa: Copy the vlan_filtering setting on the CPU port if it's global
The current behavior is not as obvious as one would assume (which is
that, if the driver set vlan_filtering_is_global = 1, then checking any
dp->vlan_filtering would yield the same result). Only the ports which
are actively enslaved into a bridge would have vlan_filtering set.

This makes it tricky for drivers to check what the global state is.
Moreover, the most obvious place to check for this setting, the CPU
port, is not populated since it's not being enslaved to the bridge.
So fix this and make the CPU port hold the global state of VLAN
filtering on this switch.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>