net: dsa: Make .port_set_egress_floods edge-triggered
The bridge port flags are kept as a bitmap by the bridge driver and set
as follows:

ip link set dev swp0 type bridge_slave flood off mcast_flood off

The command above means that the user requests unicast and multicast
flooding to be turned off.  However what DSA makes of it is that no
behavior change w.r.t. flooding is requested.

Fix that by keeping state in the dsa_port with another set of bridge
flags.  Requests to change a particular flag are detected by checking
the dsa_port bitmap against the one that just came in through the
switchdev notifier.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>