net: dsa: sja1105: Fix broken learning with vlan_filtering disabled
When put under a bridge with vlan_filtering 0, the SJA1105 ports will
flood all traffic as if learning was broken. This is because learning
interferes with the rx_vid's configured by dsa_8021q as unique pvid's.

So learning *does* work, it's just that the learnt entries never get
matched due to their unique VLAN ID.

The setting that saves the day is Shared VLAN Learning, which on this
switch family works exactly as desired: VLAN tagging still works
(untagged traffic gets the correct pvid) and FDB entries are still
populated with the correct contents including VID. Also, a frame cannot
violate the forwarding domain restrictions enforced by its classified
VLAN. It is just that the VID is ignored when looking up the FDB for
taking a forwarding decision (selecting the egress port).

The result is that frames with a learnt DMAC are no longer flooded in
the scenario described above.

Now exactly *because* SVL works as desired, we have to revisit some
earlier patches:

- It is no longer necessary to manipulate the VID of the 'bridge fdb
  {add,del}' command when vlan_filtering is off. This is because now,
  SVL in that case is enabled, so the actual VID does not matter.

- It is still desirable to hide dsa_8021q VID's in the FDB dump
  callback (show them as 1). But right now the dump callback should no
  longer hide duplicates (one per each front panel port's pvid, plus one
  for the VLAN that the CPU port is going to tag a TX frame with),
  because there shouldn't be any (the switch will match a single FDB
  entry no matter its VID anyway).

- Not really... It's no longer necessary to transform a 'bridge fdb add'
  into 5 fdb add operations, but the user might still add a fdb entry
  with any vid, and in 'bridge fdb show', all of them would appear with
  vid 1, therefore look like duplicates. So force a 'bridge fdb add' to
  insert the VID of 1, so that we can prune the duplicates at insertion
  time.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>