net: dsa: sja1105: Always configure the AVB Parameters Table
This table contains the source and destination MAC addresses used by the
switch to send follow-up metadata frames for RX timestamps.

Previously we were only configuring this table only when RX timestamping
was requested.

In SJA1105 P/Q/R/S, the PTP_CLK pin can be configured as input or output
via the CAS_MASTER field in this table. Hence it is no longer used only
for timestamping.

Because enabling RX timestamping is not actually done from this table
(but rather via SEND_META in the General Parameters Table), we can
unconditionally add the AVB Parameters Table to the static config with
no ill effect. This will simplify the configuration of the PTP_CLK pin
direction in a further patch.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>