net: dsa: sja1105: Use all management route slots for link-local xmit
The switch needs to have a temporary (one-shot) FDB entry installed each
time a frame with the DMAC of 01-80-C2-xx-xx-xx or 01-1B-19-xx-xx-xx is
sent from the CPU. In hardware there are up to 4 slots that can hold
these temporary FDB entries (management routes) at the same time.

As a result of a previous discussion with Andrew Lunn it has been
established that bad things will happen if two frames sent at the same
management DMAC are egressed concurrently on different ports, because
the switch might match them incorrectly to the slots, based on time of
arrival rather than egress port.

That is all true, but it is based on looking incorrectly at the problem.
Originally I had thought that there are 4 management routes because
there are 4 front-panel switch ports. But the real reason is that there
are 4 management routes "just because". And that is important, because
if we can affine all frames sent to the same DMAC to the same management
slot, and then serialize access to that slot only, then we can
effectively use all slots in parallel, and not just one. And of course,
a crude function that ensures DMAC-to-slot affinity is a 2-bit hash of
the DMAC.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>